fix: 워크플로우 스크립트 수정(간소화 및 네트워크 에러 수정)
